requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.0
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
idna==3.10
l18n==2021.3
laces==0.1.1
Markdown==3.7
openpyxl==3.1.5
pillow==10.4.0
pillow_heif==0.18.0
PyJWT==2.9.0
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
telepath==0.3.1
urllib3==2.2.3
wagtail==6.2.2
wheel==0.44.0
Willow==1.8.0
